Add entry for requests.codes.server_error in CONTRAIL_EXCEPTION_MAP
This patch adds the key 'requests.codes.server_error' in the
'CONTRAIL_EXCEPTION_MAP' dictionary to avoid KeyError exception
if config api server encounters an unexpected exception.
Al though the real fix should be in the config api server (as
it should not have unknown exceptions) it's better to address
it here.

Change-Id: Ifb105acf46aa00074e712c535e8cd7d7becc71b6
Closes-bug: #1432600