Use unique Id for the LBaaS service instance
When stressing the lbaas with a loop creating/deleting
vips the svc_monitor looses the references to the related
resources due to a race between the cleanup mechanism
of the svc and the service instance creation/deletion.

This patch generates an unique id for each service instance,
so that the svc is able to cleanup the related resource
properly. This unique id is stored in the KV. There is a
fallback to the old name in the clear method in order
to handle vip created before this patch.

Change-Id: I840956e9ef9a257bc1b6972efa36e9d72045dfdf