Refactor contrail neutron plugin
This patch adds a new version of contrail neutron plugin -
NeutronPluginContrailCoreV3.

This version of plugin contains the refactored vnc_openstack
code which uses the vnc_api to talk to the config api server.

The user can use either NeutronPluginContrailCoreV2 or
NeutronPluginContrailCoreV3
This patch is still in WIP. Uploaded the patch to get
feedback about the code.

Working on the test cases.

Change-Id: I47eae5c2fe0e65e8fd1c29ba4f20b8147dfe109a