Emulate format change of the project id in tests
On a classic installation the UUID format between VNC and neutron
is not the same.

Tenant ids used in the neutron tests are not UUIDs so the
_project_id_vnc_to_neutron and _project_id_neutron_to_vnc were
mocked to do nothing.

This change mimic the format change between VNC id and Neutron id
in the test cases for better testing.

Change-Id: I28d53be9e783b297903d83ab66283dfe04a2e9d0
Partial-Bug: #1484436