Allow for port migration from DPDK to kernel nodes
If you create a port and use it with a VM on a DPDK vRouter node, the port's
vif_type and vif_details fields will be updated accordingly. If you delete the
VM and decide to use the port again but now on a kernel vRouter node, the
vif_type and vif_details have to be updated to reflect the node change.

Also, this cherry-picks a fix from the R3.0 brach from the commit:
2f5560035a533a020d7a0e198f8be63fca984b6a

Change-Id: Id0cb8796a1648ee901ae2cc922dbd0d092405b34
Partial-Bug: #1544411