Allow for port migration from DPDK to kernel nodes
If you create a port and use it with a VM on a DPDK vRouter node, the port's
vif_type and vif_details fields will be updated accordingly. If you delete the
VM and decide to use the port again but now on a kernel vRouter node, the
vif_type and vif_details have to be updated to reflect the node change.

Change-Id: I0b0dc4a1458c2490993ef52c0eaad56ffca577aa
Partial-Bug: #1544411