Remove 'propagate_exc' parameter in _get_vrouter_config()
A 'propagate_exc' parameter was added to v2 plugin resources action
methods but was not implemented on the v3 plugin. Since, the v3
plugin cannot use anymore the method '_get_vrouter_config' because
it does not provide the 'propagate_exc' parameter.
That patch remove the 'propagate_exc' parameter to realign v2 and v3 on
a common plugin interface class/methods and uses the
'_raise_contrail_error' method to propagate the 'VirtualRouterNotFound'
exception.

Change-Id: I5e6afe7a817e1ff99a559598f7c8ac0ce668398c
Closes-Bug: #1561135