Monkey patch Neutron to load default service
Neutron loads a list of default services when it is initializing but most
of that services relie on the Neutron DB model. As Contrail does not use
that model, all that services cannot work.
That patch permits to load services declared under python entry point
'neutron.extension_services.contrail' if they exist and ignore them if
not (only for the default service extensions list loaded by Neutron).
It also implement a contrail TimeStamp service extension (API get with
the change_since attribute not yet implemented).

Closes-Bug: #1729052
Change-Id: I755f3570557556248cd3dc5730352cb582dfa043