Monkey patch Neutron to load default service
Neutron loads a list of default services when it initialize but most of
that services relies on the Neutron DB model. As Contrail does not use
that model, all that services cannot work.
That patch permits to load services declared under python entry point
'neutron.extension_services.contrail' if they exists and ignore them if
not (only for the default services list loaded by Neutron).
It also implement a contrail TimeStamp service extension (API get with
the change_since attribute not yet implemented).

Change-Id: I755f3570557556248cd3dc5730352cb582dfa043