Add Contrail TimeStamp service plugin
Since patch [0], Neutron does not load default service plugins if the
core driver is not base on the Neutron DB model as that services depend
on that model.
This patch implements its own service plugin 'TimeStamp' and add it to
the 'neutron.service_plugins' python namespace. To enable it, it must be
specified in a neutron config files like this:

  service_plugins = contrail-timestamp

Closes-Bug: #1729052
[0] https://review.openstack.org/#/c/518788/

Change-Id: I0aa685543427ec15cec95e4ac75754826dffbb86